requirements.txt
maturin
PySide6
pydantic
pyyaml
